make changes from review

+++ contracts/libs/UintArrayOps.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity ^0.8.4;
 
-import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/utils/math/SafeCast.sol";
 
 library UintArrayOps {
@@ -15,7 +14,7 @@ library UintArrayOps {
         return _sum;
     }
 
-    function signedDifference(uint[] memory a, uint[] memory b) internal pure returns (int[] memory _difference) {
+    function signedDifference(uint256[] memory a, uint256[] memory b) internal pure returns (int256[] memory _difference) {
         require(a.length == b.length, "Arrays must be the same length");
 
         _difference = new int256[](a.length);
@@ -27,10 +26,11 @@ library UintArrayOps {
         return _difference;
     }
 
-    function positiveDifference(uint[] memory a, uint[] memory b) internal pure returns (uint[] memory _positiveDifference) {
+    /// @dev given two int arrays a & b, returns an array c such that c[i] = a[i] - b[i], with negative values truncated to 0
+    function positiveDifference(uint256[] memory a, uint256[] memory b) internal pure returns (uint256[] memory _positiveDifference) {
         require(a.length == b.length,  "Arrays must be the same length");
 
-        _positiveDifference = new uint[](a.length);
+        _positiveDifference = new uint256[](a.length);
 
         for (uint256 i=0; i < a.length; i++) {
             if (a[i] > b[i]) {
